--- POI_UPV/nbproject/private/private.properties
@@ -1,1 +1,0 @@
-user.properties.file=C:\\Users\\clara\\AppData\\Roaming\\NetBeans\\22\\build.properties
