--- POI_UPV/src/poiupv/FXMLDocument.fxml
@@ -1,142 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.String?>
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.Menu?>
-<?import javafx.scene.control.MenuBar?>
-<?import javafx.scene.control.MenuButton?>
-<?import javafx.scene.control.MenuItem?>
-<?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.control.Slider?>
-<?import javafx.scene.control.SplitPane?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.Pane?>
-<?import javafx.scene.layout.Region?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<VBox prefHeight="556.0" prefWidth="853.0" stylesheets="@../resources/estilos.css" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="poiupv.FXMLDocumentController">
-
-  <!-- Barra superior -->
-  <VBox prefHeight="83.0" prefWidth="853.0">
-    <children>
-         <HBox prefHeight="23.0" prefWidth="853.0">
-            <children>
-          <MenuBar>
-            <menus>
-              <Menu mnemonicParsing="false" text="Ayuda">
-                <items>
-                  <MenuItem mnemonicParsing="false" onAction="#about" text="Acerca de" />
-                </items>
-              </Menu>
-            </menus>
-          </MenuBar>
-          <Button alignment="CENTER_RIGHT" contentDisplay="CENTER" mnemonicParsing="false" nodeOrientation="LEFT_TO_RIGHT" onAction="#onLogin" style="-fx-background-color: blue;" styleClass="contrast" text="Log in" textAlignment="CENTER" textFill="WHITE">
-            <HBox.margin>
-              <Insets left="550.0" />
-            </HBox.margin>
-          </Button>
-            </children>
-         </HBox>
-      <HBox alignment="CENTER_RIGHT" fillHeight="false" spacing="5.0" VBox.vgrow="NEVER">
-        <children>
-          <Region HBox.hgrow="ALWAYS" />
-               <Button fx:id="centerButton" mnemonicParsing="false" onAction="#center" text="Centrar" />
-          <Label text="Zoom" />
-          <Button mnemonicParsing="false" onAction="#zoomOut" text="-" />
-          <Slider fx:id="zoom_slider" HBox.hgrow="SOMETIMES" />
-          <Button layoutX="10.0" layoutY="10.0" mnemonicParsing="false" onAction="#zoomIn" text="+" />
-        </children>
-        <VBox.margin>
-          <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-        </VBox.margin>
-      </HBox>
-      <HBox prefHeight="44.0" prefWidth="853.0">
-        <children>
-          <Button mnemonicParsing="false" text="Button" />
-          <Button mnemonicParsing="false" text="Button" />
-          <Button mnemonicParsing="false" text="Button" />
-          <Button mnemonicParsing="false" text="Button" />
-        </children>
-      </HBox>
-    </children>
-  </VBox>
-
-  <!-- Panel dividido principal -->
-  <SplitPane fx:id="splitPane" dividerPositions="0.35" VBox.vgrow="ALWAYS">
-    <items>
-      <!-- 1) Sección de preguntas -->
-      <VBox fx:id="seccionPreguntas" alignment="CENTER" prefWidth="250.0">
-        <children>
-          <Text fx:id="tituloTest" strokeType="OUTSIDE" strokeWidth="0.0" text="Test">
-            <font>
-              <Font name="System Bold" size="20.0" />
-            </font>
-            <VBox.margin>
-              <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-            </VBox.margin>
-          </Text>
-
-          <HBox fx:id="botonesSeleccionPregunta" alignment="CENTER" spacing="10">
-            <children>
-              <Button fx:id="seleccionarPregunta" mnemonicParsing="false" onAction="#seleccionarAccion" text="Seleccionar pregunta" />
-              <Button fx:id="preguntaRandom" mnemonicParsing="false" onAction="#seleccionRandom" text="Pregunta aleatoria" />
-            </children>
-            <VBox.margin>
-              <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-            </VBox.margin>
-          </HBox>
-
-          <Text fx:id="enunciadoPregunta" strokeType="OUTSIDE" strokeWidth="0.0">
-            <font>
-              <Font size="15.0" />
-            </font>
-            <VBox.margin>
-              <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-            </VBox.margin>
-          </Text>
-        </children>
-      </VBox>
-
-      <!-- ScrollPane con la carta náutica -->
-      <ScrollPane fx:id="map_scrollpane" hbarPolicy="ALWAYS" onMouseMoved="#showPosition" onMousePressed="#addPoi" pannable="true" vbarPolicy="ALWAYS">
-        <content>
-          <Pane>
-            <children>
-              <ImageView fitHeight="5577.0" fitWidth="8674.0" layoutX="70.0" layoutY="49.0" pickOnBounds="true" preserveRatio="true">
-                <image>
-                  <Image url="@../resources/carta_nautica.jpg" />
-                </image>
-              </ImageView>
-              <MenuButton fx:id="map_pin" mnemonicParsing="false" prefHeight="60.0" prefWidth="48.0" visible="false">
-                <items>
-                  <MenuItem fx:id="pin_info" mnemonicParsing="false" text="Action 1" />
-                </items>
-                <styleClass>
-                  <String fx:value="menu-button" />
-                  <String fx:value="map-pin" />
-                </styleClass>
-              </MenuButton>
-            </children>
-          </Pane>
-        </content>
-      </ScrollPane>
-    </items>
-  </SplitPane>
-
-  <!-- Barra inferior con posición del ratón -->
-  <HBox alignment="CENTER_RIGHT">
-    <children>
-      <Label fx:id="mousePosition" text="SceneX:    SceneY:" wrapText="true" HBox.hgrow="ALWAYS" />
-    </children>
-    <VBox.margin>
-      <Insets bottom="12" right="12" top="6" />
-    </VBox.margin>
-  </HBox>
-
-</VBox>
