+++ Proyecto_IPC/nbproject/genfiles.properties
@@ -1,0 +1,5 @@
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=b609ac1b
+nbproject/build-impl.xml.script.CRC32=a2179f4d
+nbproject/build-impl.xml.stylesheet.CRC32=12e0a6c2@1.111.0.48
