+++ Proyecto_IPC/nbproject/private/private.properties
@@ -1,0 +1,1 @@
+user.properties.file=C:\\Users\\clara\\AppData\\Roaming\\NetBeans\\22\\build.properties
