+++ Proyecto_IPC/src/carta_navegacion/FXMLDisplayProblems.fxml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+
+<VBox alignment="CENTER" minHeight="-Infinity" minWidth="-Infinity" prefHeight="700.0" prefWidth="500.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="carta_navegacion.FXMLDisplayProblemsController">
+   <children>
+      <ListView fx:id="listaProblemas" prefHeight="650.0" prefWidth="480.0">
+         <VBox.margin>
+            <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+         </VBox.margin>
+      </ListView>
+      <HBox alignment="BOTTOM_RIGHT" prefHeight="30.0" prefWidth="480.0">
+         <VBox.margin>
+            <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+         </VBox.margin>
+         <children>
+            <Button fx:id="seleccionarProblema" mnemonicParsing="false" onAction="#seleccionarAccion" text="Aceptar" />
+         </children>
+      </HBox>
+   </children>
+</VBox>
